requirements.txt
itsdangerous==0.24
Werkzeug==0.10.4
WTForms==2.0.2
SQLAlchemy==1.0.9
MarkupSafe==0.23
Jinja2==2.8
Flask==0.10.1
Flask-Injector==0.6.2
Flask-SQLAlchemy==2.1
Flask-WTF==0.10
